session: retry on empty API responses, widen scroll offset to u32
Empty API responses (no content, reasoning, or tool calls) were silently treated as a completed turn, leaving the session idle with no user-visible indication of what happened. The immediate cause was a content_filter finish reason from the provider, but any empty response — provider glitches, truncated streams, spurious length stops — hit the same path. streamOneTurn now returns TransientError for empty responses instead of Done, routing them through the existing retry loop with exponential backoff (up to 5 attempts). If all retries fail, the error is surfaced to the user via LogEvent::Error rather than silent death.

Agent panel scroll fields (scrollOffset, lastScrollY, lastMaxScroll) and the related displayOffset/delta plumbing in selection.rs and app.rs were u16, capping at 65535 visual lines. Long conversations with wrapped content exceed this, causing scroll arithmetic to overflow and the view to jump or lock. All scroll-position fields and their call sites are widened to u32. Heights from Rect remain u16 and are cast at the point of use.

+++ crates/deck/src/app.rs
@@ -1927,7 +1927,7 @@ async fn runLoop(
                 selState.inputContentRect = agentPanel.lastInputRect;
 
                 // Expand selection for double/triple/quad click (needs Buffer).
-                let termOffset = termPane.activeStateRef().displayOffset() as u16;
+                let termOffset = termPane.activeStateRef().displayOffset() as u32;
                 let agentOffset = agentPanel.displayOffset();
                 if let Some((panel, clickCount)) = selState.pendingExpand.take() {
                     if panel == PanelId::Input {
@@ -4728,9 +4728,9 @@ fn handleMouse(
     deckUpdateTx: &mpsc::Sender<DeckUpdate>,
 ) -> bool {
     // Resolve display offset for the given panel.
-    fn panelOffset(panel: PanelId, termPane: &TerminalPane, agentPanel: &AgentPanel) -> u16 {
+    fn panelOffset(panel: PanelId, termPane: &TerminalPane, agentPanel: &AgentPanel) -> u32 {
         match panel {
-            PanelId::Terminal => termPane.activeStateRef().displayOffset() as u16,
+            PanelId::Terminal => termPane.activeStateRef().displayOffset() as u32,
             PanelId::Agent => agentPanel.displayOffset(),
             PanelId::Input => 0,
         }
@@ -4976,7 +4976,7 @@ fn handleMouse(
                     if selState.selectingIn == Some(PanelId::Terminal) {
                         let (_, screenRow) =
                             selState.toLocal(PanelId::Terminal, mouse.column, mouse.row);
-                        let offset = termPane.activeStateRef().displayOffset() as u16;
+                        let offset = termPane.activeStateRef().displayOffset() as u32;
                         if let Some(sel) = &mut selState.termSelection {
                             sel.extendRow(selection::toGridLine(screenRow, offset));
                         }
@@ -5012,7 +5012,7 @@ fn handleMouse(
                     if selState.selectingIn == Some(PanelId::Terminal) {
                         let (_, screenRow) =
                             selState.toLocal(PanelId::Terminal, mouse.column, mouse.row);
-                        let offset = termPane.activeStateRef().displayOffset() as u16;
+                        let offset = termPane.activeStateRef().displayOffset() as u32;
                         if let Some(sel) = &mut selState.termSelection {
                             sel.extendRow(selection::toGridLine(screenRow, offset));
                         }
@@ -5223,7 +5223,7 @@ fn extractTerminalUnwrapped(
     sel: &selection::Selection,
     area: ratatui::layout::Rect,
     buf: &ratatui::buffer::Buffer,
-    displayOffset: u16,
+    displayOffset: u32,
     termState: &TerminalState,
 ) -> String {
     if sel.isEmpty() {
